swipe-back: extract shared post-commit cleanup + header-portal ref hook
/simplify pass before PR: the commit-badge/#header-portal transform
resets + scrollbar unlock were copied across 4 destination routes
(home, GroupPageInner, PollDetailPageInner, SettingsPage) — now one
resetSwipeBackChrome() in lib/useSwipeBackGesture.ts. The 3-copy
#header-portal ref effect becomes useHeaderPortalRef(). The settings
BACKDROP instance now skips its server fetches (apiGetMe / providers /
passkeys / profile sync) — it lives under a sub-second gesture and
displays from the seeded caches; the real route re-fetches anyway.
Also drops the unused createRankedOptions import.
